% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_test_function.R
\name{calculate_cliffs_delta}
\alias{calculate_cliffs_delta}
\title{Calculate Cliff's Delta (Effect Size) for Wilcoxon Test}
\usage{
calculate_cliffs_delta(data_table, group_column, variable_column)
}
\arguments{
\item{data_table}{A data table containing the groups to compare.}

\item{group_column}{The name of the column that divides the data into two groups.}

\item{variable_column}{The name of the column for which to calculate Cliff's Delta.}
}
\value{
An object containing the estimate of Cliff's Delta and its confidence interval.
}
\description{
This function calculates Cliff's Delta, an effect size measure for the Wilcoxon test,
given two groups in a data table.
}
\examples{
dt <- data.table(group = sample(0:1, 100, replace = TRUE),
                 value = rnorm(100))
calculate_cliffs_delta(dt, "group", "value")
}
