% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_test_function.R
\name{calculate_power}
\alias{calculate_power}
\title{Calculate Power for Statistical Tests}
\usage{
calculate_power(
  data_frame,
  group_column,
  variable_column,
  stats_method,
  age_column = "Age_visit"
)
}
\arguments{
\item{data_frame}{A data frame containing the data for the power calculation.}

\item{group_column}{The name of the column that divides the data into groups.}

\item{variable_column}{The name of the column for which to calculate the power.}

\item{stats_method}{A string specifying the type of statistical test;
can be "wilcox" for Wilcoxon test, "logistic" for Logistic Regression,
or "regression" for Linear Regression.}

\item{age_column}{The name of the age column, with a default value of "Age_visit".}
}
\value{
The calculated power as a numeric value.
}
\description{
This function calculates the statistical power for different types of tests
(Wilcoxon, Logistic Regression, Linear Regression) based on simulation.
}
\examples{
df <- data.frame(group = sample(0:1, 100, replace = TRUE),
                 value = rnorm(100),
                 age = rnorm(100, mean = 50, sd = 10), Sex = sample(0:1, 100, replace = TRUE))
calculate_power(df, "group", "value", "wilcox")
calculate_power(df, "group", "value", "logistic", age_column = "age")
calculate_power(df, "group", "value", "regression", age_column = "age")
}
