% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_Check.R
\name{identify_na_columns}
\alias{identify_na_columns}
\title{Identify Columns with NA and Count Them}
\usage{
identify_na_columns(dt)
}
\arguments{
\item{dt}{A data table.}
}
\value{
A list with names of columns containing NA and the count of such columns.
}
\description{
Identify Columns with NA and Count Them
}
\examples{
dt <- data.table(a = c("apple", "", "banana"), b = c("", "orange", "grape"))
dt <- replace_empty_with_na(dt)
na_info <- identify_na_columns(dt)
print(na_info)
}
