% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plots.R
\name{quick_bar_plot}
\alias{quick_bar_plot}
\title{Quick Bar Plot}
\usage{
quick_bar_plot(data_frame, col, base_size = 12, text_size = 12)
}
\arguments{
\item{data_frame}{A data frame containing the data for plotting.}

\item{col}{The name of the column to be plotted.}

\item{base_size}{Base font size for the plot (default is 12).}

\item{text_size}{Text font size for the plot (default is 12).}
}
\value{
A ggplot2 bar plot object.
}
\description{
Generates a bar plot for a given column of a data frame using ggplot2.
}
\examples{
df <- data.frame(category = sample(c("A", "B", "C"), 100, replace = TRUE))
quick_bar_plot(df, "category")
}
