% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plots.R
\name{quick_box_plot}
\alias{quick_box_plot}
\title{Create a Box Plot}
\usage{
quick_box_plot(data_frame, x, y, title = x)
}
\arguments{
\item{data_frame}{A data frame containing the data.}

\item{x}{The name of the categorical column for the x-axis.}

\item{y}{The name of the numeric column for the y-axis.}

\item{title}{The title of the plot (defaults to the name of the x-axis column).}

\item{cut_width}{The width of the grouping for the box plot (default is 0.1).}
}
\value{
A ggplot2 box plot object.
}
\description{
Generates a box plot for comparing distributions across different categories.
}
\examples{
df <- data.frame(category = sample(c("A", "B", "C"), 100, replace = TRUE), value = rnorm(100))
quick_box_plot(df, "category", "value")
}
