% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_plots.R
\name{quick_scatter_plot}
\alias{quick_scatter_plot}
\title{Create a Scatter Plot}
\usage{
quick_scatter_plot(data_frame, x, y, title = x)
}
\arguments{
\item{data_frame}{A data frame containing the data.}

\item{x}{The name of the column for the x-axis.}

\item{y}{The name of the column for the y-axis.}

\item{title}{The title of the plot (defaults to the name of the x-axis column).}
}
\value{
A ggplot2 scatter plot object.
}
\description{
Generates a scatter plot comparing two numeric columns.
}
\examples{
df <- data.frame(x = rnorm(100), y = rnorm(100))
quick_scatter_plot(df, "x", "y")
}
