% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_summary_stats.R
\name{quick_summary_stats}
\alias{quick_summary_stats}
\title{Quick Summary of Data}
\usage{
quick_summary_stats(data_table, roi, symptom)
}
\arguments{
\item{data_table}{A data table containing the data to be summarized.}

\item{roi}{The name of the ROI column in the data table.}

\item{symptom}{The name of the symptom column in the data table.}
}
\value{
Prints summary statistics to the console.
}
\description{
This function prints a quick summary of specified ROI (Region of Interest) and symptom
statistics from the given data table.
}
\examples{
dt <- data.table("hippocampus" = c(1.5, 2.1, 1.8), "memory" = c(0, 1, 0))
quick_summary_stats(dt, "hippocampus", "memory")
}
