% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_test_function.R
\name{stats_test_function}
\alias{stats_test_function}
\title{Perform Statistical Tests on Multiple Columns}
\usage{
stats_test_function(
  data_frame,
  columns_to_test,
  reference_column,
  stats_method,
  Age = "Age_visit",
  use_weights = FALSE
)
}
\arguments{
\item{data_frame}{A data frame containing the data for the tests.}

\item{columns_to_test}{A vector of column names to be tested.}

\item{reference_column}{The name of the reference column for comparison in tests.}

\item{stats_method}{A string specifying the type of test to perform;
can be "wilcox" for Wilcoxon test, "logistic" for Logistic Regression,
or "regression" for Linear Regression.}

\item{Age}{The name of the age column, with a default value of "Age_visit".}

\item{use_weights}{A boolean indicating whether to use weights in logistic regression.}
}
\value{
A data frame with test results for each column, including p-values and effect sizes.
}
\description{
This function applies specified statistical tests to multiple columns of a data frame
and returns the results, including p-values and effect sizes.
}
\examples{
df <- data.frame(group = sample(0:1, 100, replace = TRUE),
                  counts = sample(0:20, 100, replace = TRUE),
                 value1 = rnorm(100),
                 value2 = rnorm(100),
                 age = rnorm(100, mean = 50, sd = 10),
                 Sex = sample(c("Male", "Female"), 100, replace = TRUE))
setDT(df)
stats_test_function(df, c("value1", "value2"), "group", "wilcox")
stats_test_function(df, c("value1", "value2"), "group", "logistic", Age = "age")
stats_test_function(df, c("value1", "value2"), "counts", "regression", Age = "age")
}
